moved CompFiles struct and method to its own module
